Delete redundant code and simplify files
